+++ src/edit/loader/loaderGLTF.ts
@@ -1,6 +1,6 @@
 import { GLTFLoader, GLTF } from "three/addons/loaders/GLTFLoader.js";
 import { DRACOLoader } from "three/addons/loaders/DRACOLoader.js";
-import { AnimationMixer } from "three";
+// import { AnimationMixer } from "three";
 
 let gltfLoader: null | GLTFLoader = null;
 let dracoLoader: null | DRACOLoader = null;